Fixed build script
Build script was crashing if dist and build folders didn't exist

+++ build.py
@@ -1,9 +1,13 @@
 from os import mkdir,system
+from os.path import exists, isdir
 from shutil import copy2,copytree,rmtree
 
 # Cleaning folders
-rmtree("dist")
-rmtree("build")
+if exists("dist") and isdir("dist"):
+  rmtree("dist")
+
+if exists("build") and isdir("build"):
+  rmtree("build")
 
 # Creating folders
 mkdir("dist")
@@ -11,7 +15,6 @@ mkdir("build")
 
 # Populating dist
 copytree("src/icons/","dist/icons")
-copy2("src/manifest.webmanifest", "dist")
 
 # Populating build
 copytree("src/libs","build/libs")
@@ -24,3 +27,4 @@ system("cd build && emcmake cmake . -DCMAKE_BUILD_TYPE=Release && cmake --build 
 
 # Minify html
 system("html-minifier --collapse-whitespace --remove-comments --remove-optional-tags --remove-redundant-attributes --remove-script-type-attributes --remove-tag-whitespace --use-short-doctype --minify-css true --minify-js true src/index.html -o dist/index.html")
+system("html-minifier --collapse-whitespace --remove-comments --remove-optional-tags --remove-redundant-attributes --remove-script-type-attributes --remove-tag-whitespace --use-short-doctype --minify-css true --minify-js true src/manifest.webmanifest -o dist/manifest.webmanifest")
